feat(BE-24): complete mine service upload endpoint
