+++ pkg/repository/storage/mongodb/mongo.go
@@ -3,9 +3,11 @@ package mongodb
 import (
 	"context"
 	"fmt"
+	"log"
+
 	"github.com/workshopapps/pictureminer.api/internal/constants"
 	"github.com/workshopapps/pictureminer.api/internal/model"
-	"log"
+	"go.mongodb.org/mongo-driver/bson"
 
 	"github.com/workshopapps/pictureminer.api/internal/config"
 	"github.com/workshopapps/pictureminer.api/utility"
@@ -16,18 +18,19 @@ import (
 
 var (
 	ctx         context.Context
-	mongoclient *mongo.Client
+	mongoClient *mongo.Client
 )
 
 func Connection() (db *mongo.Client) {
-	return mongoclient
+	return mongoClient
 }
 
 func ConnectToDB() *mongo.Client {
+	var err error
 	logger := utility.NewLogger()
 	uri := config.GetConfig().Mongodb.Url
 	mongo_connection := options.Client().ApplyURI(uri)
-	mongoClient, err := mongo.Connect(ctx, mongo_connection)
+	mongoClient, err = mongo.Connect(ctx, mongo_connection)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -42,7 +45,6 @@ func ConnectToDB() *mongo.Client {
 	fmt.Println("MONGO CONNECTION ESTABLISHED")
 	logger.Info("MONGO CONNECTION ESTABLISHED")
 
-	mongoclient = mongoClient
 	return mongoClient
 }
 
@@ -55,8 +57,63 @@ func GetCollection(client *mongo.Client, databaseName, collectionName string) *m
 //creating a user in the data base
 
 func CreateUser(context context.Context, User model.UserStruct) {
-	mongoClient := Connection()
 	imageDB := mongoClient.Database(constants.UserDatabase)
 	userCollection := imageDB.Collection(constants.UserDatabase)
 	userCollection.InsertOne(context, User)
+}
+
+func MongoPost(collection string, data interface{}) (*mongo.InsertOneResult, error) {
+	c := getCollection(collection)
+
+	result, err := c.InsertOne(ctx, data)
+	if err != nil {
+		return nil, err
+	}
+
+	return result, nil
+}
+
+func MongoGet(collection string, filter map[string]interface{}) ([]interface{}, error) {
+	c := getCollection(collection)
+
+	f := make(bson.M, 0)
+
+	if len(filter) == 1 {
+		for k, v := range filter {
+			f = bson.M{k: v}
+		}
+	} else if len(filter) > 1 {
+		tf := make([]bson.M, 0, len(filter))
+		for k, v := range filter {
+			tf = append(tf, bson.M{k: v})
+		}
+
+		f = bson.M{"$and": tf}
+	}
+
+	cursor, err := c.Find(ctx, f)
+	if err != nil {
+		return nil, err
+	}
+
+	results := make([]interface{}, 0)
+
+	for cursor.Next(ctx) {
+		var result interface{}
+		err := cursor.Decode(&result)
+		if err != nil {
+			return nil, err
+		}
+		results = append(results, result)
+	}
+
+	return results, nil
+}
+
+func getCollection(collection string) *mongo.Collection {
+	databaseName := config.GetConfig().Mongodb.Database
+	database := mongoClient.Database(databaseName)
+	c := database.Collection(collection)
+
+	return c
 }
